requirements.txt
line-bot-sdk>=3.0.0
